moved api request to context instead of local comp
